% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukTownData.R
\name{display_bua_buasd_status}
\alias{display_bua_buasd_status}
\title{Display BUA or BUASD status based on town name}
\usage{
display_bua_buasd_status(data, town_name)
}
\arguments{
\item{data}{The dataset containing town information.}

\item{town_name}{The name of the town to check.}
}
\value{
A string indicating whether the town is a BUA, BUASD, or neither.
}
\description{
This function checks if a town is classified as BUA or BUASD using the \code{is_bua} column.
}
