% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_bua_buasd_towns}
\alias{list_bua_buasd_towns}
\title{List of BUA and BUASD towns}
\usage{
list_bua_buasd_towns(data)
}
\arguments{
\item{data}{The dataset containing town information.}
}
\value{
A list with two data frames: BUA towns and BUASD towns.
}
\description{
This function returns a list containing two data frames: one for towns classified as BUA
and another for towns classified as BUASD.
}
