% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukTownData.R
\name{retrieve_town_data}
\alias{retrieve_town_data}
\title{Retrieve data for a specific town}
\usage{
retrieve_town_data(data, town_name)
}
\arguments{
\item{data}{The dataset to subset.}

\item{town_name}{The name of the town to retrieve.}
}
\value{
A dataframe with data for the specified town.
}
\description{
This function returns all rows related to a specific town.
}
