% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_app}
\alias{run_app}
\title{Launch the UkTownMetrics Shiny Application}
\usage{
run_app()
}
\value{
Runs the Shiny application.
}
\description{
This function launches the Shiny application included in the \code{ausbirthplace} package.
}
\examples{
\dontrun{
run_app()
}
}
